refactor: mise en place des imports exports des modules js
