+++ css/photographers.css
@@ -1,14 +1,3 @@
-@import url("https://fonts.googleapis.com/css2?family=DM+Sans:wght@400;500;700&display=swap");
-
-* {
-    box-sizing: border-box;
-    font-family: 'DM Sans', sans-serif
-}
-
-html {
-    scroll-behavior: smooth
-}
-
 /* HEADER */
 header {
     align-items: center;
