+++ css/styles.css
@@ -1,0 +1,15 @@
+@import url("https://fonts.googleapis.com/css2?family=DM+Sans:wght@400;500;700&display=swap");
+
+* {
+    box-sizing: border-box;
+    font-family: 'DM Sans', sans-serif
+}
+
+html {
+    scroll-behavior: smooth;
+}
+
+body {
+    margin-left: 2rem;
+    margin-right: 2rem;
+}
