+++ index.html
@@ -5,10 +5,9 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>FishEye - Accueil</title>
-    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.15.1/css/all.css">
-    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300&display=swap" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css2?family=Open+Sans:wght@600&display=swap" rel="stylesheet">
+    <link rel="stylesheet" href="css/styles.css">
     <link rel="stylesheet" href="css/homepage.css">
+    <script type="module" src="./js/main.js"></script>
 </head>
 
 <body>
@@ -49,10 +48,9 @@
     <a id="main-photographers_link" href="#">Passer au contenu</a>
     <main>
         <!-- PHOTOGRAPHERS -->
-        <section class="photographers" id="photographers" aria-label="photographers">
+        <section id="photographers" aria-label="photographers">
         </section>
     </main>
-    <script type="module" src="js/homepage.js"></script>
 
 </body>
