+++ js/main.js
@@ -1,0 +1,9 @@
+'use strict';
+////////////////////////////////////////////////////////////
+
+import getApiFishEye from './getApiFishEye.js'
+import scrollButton from './scrollButton.js';
+
+
+getApiFishEye();
+scrollButton();
